Reflection_Model-main/requirements.txt
transformers>=4.38.0
torch>=2.1.0
sentencepiece>=0.1.99
datasets>=2.16.0
accelerate>=0.27.2

scikit-learn>=1.4.1
evaluate>=0.4.2
sentence-transformers>=2.2.2

pandas>=2.1.4
numpy>=1.26.4

streamlit>=1.33.0
tqdm>=4.66.2
rich>=13.7.0
colorama>=0.4.6
tabulate>=0.9.0

httpx>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.1

groq>=0.4.2

tokenizers>=0.15.0
regex>=2023.12.25

rich
sentencepiece
gradio